Fixes for resize bug that showed up in demo
