+++ src/core_plugins/metrics/public/visualizations/lib/calc_dimensions.js
@@ -1,0 +1,5 @@
+export default function calcDimensions(el, scale) {
+  const newWidth = Math.floor(el.clientWidth * scale);
+  const newHeight = Math.floor(el.clientHeight * scale);
+  return [newWidth, newHeight];
+}
